requirements.txt
flask
werkzeug
azure-storage-blob
openai-whisper
torch


